feat: redis-separator
https://github.com/andrewthetechie/pydantic-aioredis/issues/192
https://github.com/andrewthetechie/pydantic-aioredis/issues/277
This adds the ability to customize the redis separator and changes the
default separator from _%&_ to :.

Additionally, it adds a way for a user to add a prefix to the key space
we use, to further differentiate their keys.

Finally, it fixes a bug with testing where redislite would not shutdown
properly and updates the fixtures to work with the newest version of
pytest_asyncio.

BREAKING CHANGE: This will result in "data loss" for existing models
stored in redis due to the change in default separator. To maintain
backwards compatbility with 0.7.0 and below, you will need to modify
your existing models to set _redis_separator = "_%&_" as a field on
them.