feat: update on setattr
https://github.com/andrewthetechie/pydantic-aioredis/issues/279
BREAKING CHANGE: This is a breaking change to how updates for model
attributes are saved to Redis. It removes the update classmethod and
replaces with with a save method on each model.

Additionally, if _auto_sync is set to true on a model (the default), it
will save to redis on any attribute change