ci: tests without redislite (#389)
* ci: tests without redislite

Thanks to @gtmanfred for invaluable debugging help and suggestions on
how to improve testing.

Redislite worked well, but it spun up an actual redis server for each
test. This caused a bunch of flakiness.

This PR moves to using fakeredis, which mocks redis entirely in python
without spinning up an external redis service. Tests are faster and more
reliable and use less system resources

* docs: fix readme layout